feat: Add scrapable check on adding Bookmark (#239)
Issue number: resolves #223

The scrapable value from the add bookmark request is checked inside
addBookmark method of BookmarkService, in order to avoid scraping the page
using JSoup and taking screenshot when the value is set to false.

This, however, do not have any impact on the import functionality - the default
behavior of scrapping and taking screenshot is kept as-is.

The unit test for addBookmark has been modified, adding a second test case
which simulates the case where scrapable is set to false. Upon receiving the
response, it is checked if the screenshot URL is empty, which implies that the
functionality worked as expected.