feat: Update SonarQube plugin to ver 5.1.0.4882 (#258)
Issue number: resolves #213

The previously-used version of the SonarQube plugin issued a warning notice,
informing about a feature getting deprecated in the next version, whenever a gradle
task was invoked.

A potential fix was to set the gradle.skipCompile flag to "true" upon calling a gradle
function, but this is not ideal. The version of the plugin was updated entirely
to avoid the warning notice, as well as to keep the plugins up-to-date in the
application.